Created Object Pooler and Fixed Sprint Regen Float Value
The sprint regen text is not a float anymore. It's an int!
Created an object pooler to spawn obstacles for player 1.

To Do:
-Create Death Wall Script to handle collisions with player and obstacle
-Player will handle obstacle
-Make obstacle despawn when hit with death wall (do in death wall
script)
-Obstacles not showing on top of background

+++ Assets/Scripts/Minigames/DragonGame/BarScript.cs
@@ -21,7 +21,7 @@ public class BarScript : MonoBehaviour {
             //because its not always gonna be health before the :, 
             //so we take out everything before the colon
             string[] temp = valueText.text.Split(':');
-            valueText.text = temp[0] + ": " + value; //make sure its lowercase v
+            valueText.text = temp[0] + ": " + (int) value; //make sure its lowercase v
             fillAmount = Map(value, 0, MaxValue, 0, 1);
         }
     }
